% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_analyte_codes.R
\docType{data}
\name{epa_analyte_codes}
\alias{epa_analyte_codes}
\title{EPA AirData Analyte Codes}
\format{
A tibble with the following columns:
\describe{
\item{analyte}{The pollutant or air quality measure extracted from the filename.}
}
}
\source{
\url{https://aqs.epa.gov/aqsweb/airdata/download_files.html}
}
\usage{
epa_analyte_codes
}
\description{
This dataset contains the queryable names for EPA analytes from the EPA AirData website.
}
\examples{
data(epa_analyte_codes)
epa_analyte_codes
}
\keyword{datasets}
