% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epa_airdata_analyte_codes.R
\name{get_epa_airdata_analyte_codes}
\alias{get_epa_airdata_analyte_codes}
\title{Scrape Table Headers from EPA AirData Webpage}
\usage{
get_epa_airdata_analyte_codes()
}
\value{
A tibble with a single column \code{label} containing unique table headers (excluding \code{"Year"}).
}
\description{
This function scrapes all table headers from the EPA AirData download page, removes the header \code{"Year"},
and returns a tibble.
}
\examples{
analyte_codes <- get_epa_airdata_analyte_codes()
print(analyte_codes)
}
