% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epa_airdata_zip_links.R
\name{get_epa_airdata_zip_links}
\alias{get_epa_airdata_zip_links}
\title{Scrape EPA AirData ZIP File Links and Metadata}
\usage{
get_epa_airdata_zip_links(archive = FALSE, archive_id = "20250126115248")
}
\arguments{
\item{archive}{Logical. If \code{TRUE}, the function retrieves data from the Wayback Machine
(Internet Archive) rather than the live EPA AirData website. Defaults to \code{FALSE}.}

\item{archive_id}{Character. The timestamp ID for the archived version of the EPA AirData website
on the Wayback Machine (only used if \code{archive = TRUE}). Defaults to \code{"20250126115248"}.}
}
\value{
A tibble with structured metadata of available ZIP files.
}
\description{
This function scrapes the EPA AirData website for downloadable ZIP file links,
extracts relevant metadata from filenames (such as year, unit of analysis, and analyte),
and returns a tidy tibble with structured information.
}
