% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epa_airdata_analyte_codes.R
\name{get_epa_airdata_analyte_codes}
\alias{get_epa_airdata_analyte_codes}
\title{Retrieve Unique Analyte Codes from EPA AirData}
\usage{
get_epa_airdata_analyte_codes()
}
\value{
A character vector containing unique analyte codes.
}
\description{
This function extracts unique analyte codes from the available EPA AirData ZIP file links.
These analyte codes serve as identifiers for different air quality measurements in the dataset.
}
\examples{
analyte_codes <- get_epa_airdata_analyte_codes()
print(analyte_codes)
}
