% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epa_airdata_analyte_tibble.R
\name{get_epa_airdata_analyte_tibble}
\alias{get_epa_airdata_analyte_tibble}
\title{Retrieve Unique Analyte Codes from Package Dataset}
\usage{
get_epa_airdata_analyte_tibble()
}
\value{
A tibble containing a single column \code{analyte} containing unique analyte codes.
}
\description{
This function extracts unique analyte codes from the available EPA AirData ZIP file links, as stored within the package.
}
\examples{
analyte_codes <- get_epa_airdata_analyte_tibble()
print(analyte_codes)
}
